+++ liquid/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 # pylint: disable=useless-import-alias,missing-module-docstring
 
-__version__ = "1.4.1"
+__version__ = "1.4.2"
 
 try:
     from markupsafe import escape as escape