+++ liquid/builtin/filters/misc.py
@@ -49,6 +49,10 @@ def default(obj: Any, default_: object = "", *, allow_false: bool = False) -> An
     if hasattr(obj, "__liquid__"):
         _obj = obj.__liquid__()
 
+    # Liquid zero is not falsy.
+    if isinstance(_obj, int) and not isinstance(_obj, bool):
+        return obj
+
     if allow_false is True and _obj is False:
         return obj